Accept is now private for Character.
Introduced a new constructor for character where we actually build a character based off of the stats given to us by the user.

Delaying construction this way makes the code far cleaner and easier to maintain. Plus the Character object itself does not need to be aware of how the data was given to it. It also means that we do not have to keep purging and clearing the object. Instead we can construct the necessary parts and then construct the object off of that.